requirements.txt
Django==5.1
djangorestframework==3.14
psycopg[binary]==3.2.9
redis==4.5
django-tenants==3.7.0
celery==5.3
django-cors-headers==4.0
django-jazzmin==2.6
pytest>=8.3,<9
pytest-django>=4.8,<5
